+++ ntex-tls/examples/simple-client.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use ntex::{Pipeline, codec, connect, util::Bytes, util::Either};
+use ntex::{Pipeline, SharedCfg, codec, connect, util::Bytes, util::Either};
 use tls_openssl::ssl::{self, SslFiletype, SslMethod, SslVerifyMode};
 
 #[ntex::main]
@@ -20,7 +20,10 @@ async fn main() -> io::Result<()> {
     builder.set_verify(SslVerifyMode::NONE);
 
     // openssl connector
-    let connector = Pipeline::new(connect::openssl::SslConnector::new(builder.build()));
+    let connector = Pipeline::with_st(
+        SharedCfg::default(),
+        connect::openssl::SslConnector::new(builder.build()),
+    );
 
     let io = connector.call("127.0.0.1:8443".into()).await.unwrap();
     println!("Connected to ssl server");