+++ ntex/src/http/test.rs
@@ -354,6 +354,7 @@ impl TestServer {
         }
     }
 
+    #[must_use]
     /// Set client timeout
     pub fn set_client_timeout(mut self, timeout: Seconds, connect_timeout: Millis) -> Self {
         self.cfg = SharedCfg::new("TEST-CLIENT")
@@ -480,12 +481,12 @@ impl TestServer {
             .set_alpn_protos(b"\x08http/1.1")
             .map_err(|e| log::error!("Cannot set alpn protocol: {e:?}"));
 
-        WsClient::builder(self.url(path), self.cfg.clone())
+        WsClient::builder(self.url(path))
             .address(self.addr)
             .timeout(Seconds(30))
             .openssl(builder.build())
             .take()
-            .build()
+            .build(self.cfg.clone())
             .unwrap()
             .connect()
             .await