+++ ntex/src/ws/client.rs
@@ -802,14 +802,14 @@ mod tests {
 
     #[crate::rt_test]
     async fn test_debug() {
-        let mut builder = WsClient::builder("http://localhost", SharedCfg::default())
+        let mut builder = WsClient::builder("http://localhost")
             .header("x-test", "111")
             .take();
         let repr = format!("{builder:?}");
         assert!(repr.contains("WsClientBuilder"));
         assert!(repr.contains("x-test"));
 
-        let client = builder.build().unwrap();
+        let client = builder.build(SharedCfg::default()).unwrap();
         let repr = format!("{client:?}");
         assert!(repr.contains("WsClient"));
         assert!(repr.contains("x-test"));
@@ -817,10 +817,10 @@ mod tests {
 
     #[crate::rt_test]
     async fn header_override() {
-        let req = WsClient::builder("http://localhost", SharedCfg::default())
+        let req = WsClient::builder("http://localhost")
             .header(header::CONTENT_TYPE, "111")
             .set_header(header::CONTENT_TYPE, "222")
-            .build()
+            .build(SharedCfg::default())
             .unwrap();
 
         assert_eq!(
@@ -836,18 +836,18 @@ mod tests {
 
     #[crate::rt_test]
     async fn basic_errs() {
-        let err = WsClient::builder("localhost", SharedCfg::default())
-            .build()
+        let err = WsClient::builder("localhost")
+            .build(SharedCfg::default())
             .err()
             .unwrap();
         assert!(matches!(err, WsClientBuilderError::MissingScheme));
-        let err = WsClient::builder("unknown://localhost", SharedCfg::default())
-            .build()
+        let err = WsClient::builder("unknown://localhost")
+            .build(SharedCfg::default())
             .err()
             .unwrap();
         assert!(matches!(err, WsClientBuilderError::UnknownScheme));
-        let err = WsClient::builder("/", SharedCfg::default())
-            .build()
+        let err = WsClient::builder("/")
+            .build(SharedCfg::default())
             .err()
             .unwrap();
         assert!(matches!(err, WsClientBuilderError::MissingHost));
@@ -855,9 +855,9 @@ mod tests {
 
     #[crate::rt_test]
     async fn basic_auth() {
-        let client = WsClient::builder("http://localhost", SharedCfg::default())
+        let client = WsClient::builder("http://localhost")
             .basic_auth("username", Some("password"))
-            .build()
+            .build(SharedCfg::default())
             .unwrap();
 
         assert_eq!(
@@ -871,9 +871,9 @@ mod tests {
             "Basic dXNlcm5hbWU6cGFzc3dvcmQ="
         );
 
-        let client = WsClient::builder("http://localhost", SharedCfg::default())
+        let client = WsClient::builder("http://localhost")
             .basic_auth("username", None)
-            .build()
+            .build(SharedCfg::default())
             .unwrap();
 
         assert_eq!(
@@ -904,9 +904,9 @@ mod tests {
 
     #[crate::rt_test]
     async fn bearer_auth() {
-        let client = WsClient::builder("http://localhost", SharedCfg::default())
+        let client = WsClient::builder("http://localhost")
             .bearer_auth("someS3cr3tAutht0k3n")
-            .build()
+            .build(SharedCfg::default())
             .unwrap();
 
         assert_eq!(
@@ -940,7 +940,7 @@ mod tests {
     #[cfg(feature = "cookie")]
     #[crate::rt_test]
     async fn basics() {
-        let mut builder = WsClient::builder("http://localhost/", SharedCfg::default())
+        let mut builder = WsClient::builder("http://localhost/")
             .origin("test-origin")
             .max_frame_size(100)
             .server_mode()
@@ -957,7 +957,7 @@ mod tests {
         assert!(builder.inner.as_ref().unwrap().server_mode);
         assert_eq!(builder.protocols, Some("v1,v2".to_string()));
 
-        let client = builder.build().unwrap();
+        let client = builder.build(SharedCfg::default()).unwrap();
         assert_eq!(
             client.head.headers.get(header::CONTENT_TYPE).unwrap(),
             header::HeaderValue::from_static("json")
@@ -965,19 +965,15 @@ mod tests {
 
         let _ = client.connect().await;
 
+        assert!(WsClient::builder("/",).build(SharedCfg::default()).is_err());
         assert!(
-            WsClient::builder("/", SharedCfg::default())
-                .build()
+            WsClient::builder("http:///test",)
+                .build(SharedCfg::default())
                 .is_err()
         );
         assert!(
-            WsClient::builder("http:///test", SharedCfg::default())
-                .build()
-                .is_err()
-        );
-        assert!(
-            WsClient::builder("hmm://test.com/", SharedCfg::default())
-                .build()
+            WsClient::builder("hmm://test.com/",)
+                .build(SharedCfg::default())
                 .is_err()
         );
     }
@@ -991,8 +987,8 @@ mod tests {
         head.method = Method::POST;
         drop(head);
 
-        let client = WsClient::builder("http://localhost", SharedCfg::default())
-            .build()
+        let client = WsClient::builder("http://localhost")
+            .build(SharedCfg::default())
             .unwrap();
         assert_eq!(client.head.method, Method::GET);
     }